+++ llama_stack/providers/inline/huggingface/datasetio/benchmarks/__init__.py
@@ -3,3 +3,4 @@
 #
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
+from .llamastack_mmlu import llamastack_mmlu  # noqa: F401
